requirements.txt
discord.py
datetime
asyncio
random
time